pyproject.toml
[project]
name = "itd-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "itd-sdk==2.0.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "sqlalchemy>=2.0.49",
]

[tool.uv.sources]
itd-sdk = { git = "https://github.com/itd-sdk/itd-sdk" }
